api_links DELETE checks canDelete and ignores if false

+++ components/tools/OmeroWeb/omeroweb/webclient/views.py
@@ -978,7 +978,10 @@ def _api_links_DELETE(conn, json_data):
                 if objLnks is None:
                     continue
                 linkType, links = objLnks
-                linkIds = [r.id.val for r in links]
+                linkIds = [r.id.val for r in links if
+                           r.getDetails().permissions.canDelete()]
+                if len(linkIds) == 0:
+                    continue
                 logger.info("api_link: Deleting %s links" % len(linkIds))
                 conn.deleteObjects(linkType, linkIds, wait=True)
                 # webclient needs to know what is orphaned
